CAMEL-24601 - fix flakiness in SmbComponentConnectionIT
one route was potentially messing with the others. I tried to use some
specific searchPattern or exclude parameter without success; The easiest
is to separate in 2 test classes and Isolate them.
Before I got a failure rate around 1 out of 5; now I got 60 in success.
